requirements.txt
# You can optionally pin to a version of the Runway Model SDK like: "runway-python==0.1.0"
runway-python
Pillow==6.2.2
dlib
face_recognition
